Clean up FFE checks of OMP directives and MP-lowering warnings
* Fix semantic checking of OMP directive placement (Fortran).  This refines
the way in which the Fortarn front-end checks the placement of OMP directives
with respect to eachother. In particular, more checks are added to prevent
work-sharing constructs from being closely nested within eachother or within
explicit tasks. This also accounts for nested directives to more accurately
track what the current "directive state" is.

* Fix MP-lowering warnings for writing to shared variables.  The MP-lowering
phase was giving spurious warnings about writing to a variables in a parallel
region that had implicit shared scope. Fixed it so that warnings aren't given
for writes in a single or critical region. Same for shared variables in a task
region that occurs lexically within a master, single, or critical section.

Contributor: Deepak

git-svn-id: http://svn.open64.net/svnroot/open64@4042 ead328b1-e903-0410-a19d-eacee4b031f6