[CAF] check to use external linker when compiling CAF programs
By default, OpenUH will use its internal gcc to link programs. When compiling
CAF programs, it will now check if the OPENUH_CAF_EXT_LINKER environment
variable is set, and if so use its value as the linker. uhcaf will set
OPENUH_CAF_EXT_LINKER to $GASNET_LD if the communication layer is GASNet and
the conduit is not SMP or UDP.

Other changes:

- Support gemini conduit in libcaf/gasnet makefile

- Fixing some bugs in non-blocking armci handle management.  Need to
  pre-allocate an array of non-blocking ARMCI handles of length nb_xfer_limit.
  The handles were not being allocated correctly before.

- A CAF runtime build will now install both static libraries and dynamic
  shared libraries.

- Tweak for contiguous access check in compiler.  This change allows
  recognition of certain contiguous accesses of remote allocatable coarray

- Updating patch level to 3.0.30.

Contributor: Deepak

git-svn-id: http://svn.open64.net/svnroot/open64@4066 ead328b1-e903-0410-a19d-eacee4b031f6