[CAF] Fix problems in normal and error termination
This commit fixes some of the oustanding issues for normal and error
termination.

Normal Termination:

When an image begins normal termination, it will send notifications to all
other images that it has stopped. If another image attempts to synchronize
with it (using a synchronization statement or allocate/deallocate statement),
then it will either go into error termination or (if a status variable is
supplied) it will return with STAT_STOPPED_IMAGE status.

In order to detect if a particular image has stopped without requiring a
remote access, every image now maintains the status of every other image
locally. The increase in memory requirement here (NUM_IMAGES() additional
bytes) is I think warranted.

TODO: If an image enters a full barrier prior to receiving the notification of
a stopped image, it will currently not detect this during the execution of the
barrier. It isn't as of yet clear whether its necessary to do this. If so,
then the barrier needs to be implemented explicitly in the CAF runtime rather
than through invocation of a barrier operation provided by one of the
underyling communication layers.

Error Termination:

When an image begins error termination and it hasn't received an error
termination notification from any other image, then it will send out these
notifications itself to all images. I've scattered many checks for this error
termination notification throughout the runtime code so that images are
frequently checking whether they should abort while executing in the runtime.
Hopefully the overhead of these checks is not noticeable.

The commit also does the following:

* Adds status variable support for the coarray allocate and deallocate
  statements

* Modifies the values of some status codes defined in iso_fortran_env

* Fixes translation for lock/unlock and event statements.  If there is
  function call in a lock/unlock or event post, event wait, event query
  statement, the front-end gave an error. For example, this was not handled
  properly:

    lock ( lk[this_image()+1] )

  This commit fixes the problem (the subroutine call these statements got
  translated to was not inserted in the right position in the statement
  table).

* Adding description of new environment variables into cafrun script

* Updating description of image-heap setting

* Fixing bug where a 0 status is return by ERROR STOP

Contributor: Deepak

git-svn-id: http://svn.open64.net/svnroot/open64@4086 ead328b1-e903-0410-a19d-eacee4b031f6