[CAF] ignore image duplicates in sync images
The spec says that there should not be duplicates in the image-list argument
for sync images. Cray's compiler will send repeated sync notifications for all
all items in the image-list, including duplicates. Our's was doing the same
for the "counter" sync images implementation, but it won't work for the ping_pong
or sense_reversing algorithms. So the alternative is we need to detect
repeated images in the image-list and either issue an error or skip over them
during the execution of sync images. I chose to do the latter, mainly because
there are some codes (e.g. EPCC's CAF microbenchmarks) that for whatever
reason have repeated items in the image-list.

Also, this scraps the old (broken) sync images implementation that was done in
ARMCI.

git-svn-id: http://svn.open64.net/svnroot/open64@4088 ead328b1-e903-0410-a19d-eacee4b031f6