[CAF] support initialized coarrays or targets and more
This commits supports save coarrays and targets which include initializers on
their declaration statements. Save coarrays and targets are allocated at the
top of the RMA segment by the runtime at the beginning of the program.
Consequentially, we need to copy over any initialization values into the RMA
segment at program startup for save coarrays and targets. The approach taken
here is to retain the initial data layout of initialized save coarrays and
targets in the .data section, and then create another global pointer symbol
which points to these values. Then, the runtime just does a memcpy of the
values from the .data section into the RMA segment.

This does require that the initialized values occupy memory space that's never
really used during the program. Ideally, it would be better if we could just
register/pin the .data section of the program, but this does not appear to be
currently supported with GASNet or ARMCI. Another option could be to keep the
initialized save coarray and targets in .data and not try to allocate them
into the RMA segment. We could support this with GASNet using Active Messages
(with performance penalty), but not in ARMCI.

Additionally, this commit fixes the translation for an array assignment to a
co-indexed struct (coarray of derived type).

Updating patch level to 3.0.35

Contributor: Deepak

git-svn-id: http://svn.open64.net/svnroot/open64@4095 ead328b1-e903-0410-a19d-eacee4b031f6