[CAF] improve output for communication stats + fix
This cleans up the output for communication stats (emitted when profiling
is enabled and --stats is specified). Also, the pid of image 1 is appended to
the file name.

This also corrects the GASNet implementation for atomic_ref, by calling to
gasnet_AMpoll. This is necessary, to detect updates made through an active
message.

updating ChangeLog for 3.0.36

git-svn-id: http://svn.open64.net/svnroot/open64@4101 ead328b1-e903-0410-a19d-eacee4b031f6