[CAF] adds node-aware algorithms for barrier and reductions
For each team, the runtime maintains a list of images that belong to an
image's intra-node set, and also a list of leader images in the team.  Using
this, we implement alternative algorithms for barriers and reductions that are
node-aware or 2-level, which may be optionally used. The 2-level barriers may
be selected via an environment variable, and it will be used only within a
team block (otherwise, the barrier provided by the underlying runtime is
used). The 2-level reduction may also be selected with an environment
variable.

Contributors: Shiyao, Deepak, Dounia

git-svn-id: http://svn.open64.net/svnroot/open64@4120 ead328b1-e903-0410-a19d-eacee4b031f6