[CAF] support CO_REDUCE, 2-level broadcast, plus other stuff
This provides some initial support for CO_REDUCE, or a user-defined reduction
in CAF, based on the current draft for TS 18508 (new parallel features).
Currently it assumes that the arguments are passed by reference, and there is
still no front-end checking that the function is valid.

A 2-level broadcast implementation. It is selected with
UHCAF_COLLECTIVES_2LEVEL, which also enables the 2-level reduction
implementation. Requires more evaluation to see if the current approach is
actually good for performance.

We also add a fix for bad initialization of use associated function result.
This is necessary when there is a module which uses another module that
contains a function interface which returns a type that has initialized
components.

We rename coarray_allocatable_allocate_new_ to coarray_allocatable_allocate_,
and get rid of the old allocatable_allocate routine.  Also, this fixes a bug
in _DEALLOCATE where it would enter into libcaf deallocate routine even for
arrays of size 0, resulting in a fatal error. Now, we check for this.

Additionally:

- removes some of the buggy or old barrier algorithm implementations,
  and keeps just 3 that are known to work: a dissemination algorithm, and two
  2-level algorithms.

- renames an environment variable UHCAF_SYNC_ALL_ALGORITHM to
  UHCAF_TEAM_BARRIER_ALGORITHM

- adds --init-team-heap option to cafrun for specifying heap size for initial
  team

- adds --with-default-caf-initial-team-heap-size configure option for
  compiler (default is 20M)

- updates the cafrun --help-env output to show newly added environment
  variables available for our implementation

Contributor: Deepak

git-svn-id: http://svn.open64.net/svnroot/open64@4123 ead328b1-e903-0410-a19d-eacee4b031f6