[CAF] fixing reuse of empty slots in RMA allocator
We encountered a problem with our RMA memory allocator scheme. Suppose p and q
are to be allocated out of the asymmetric space in our heap, like so:

allocate(p(..))
allocate(q(..))

do ...

    deallocate(p)
    allocate(p(..))

    deallocate(q)
    allocate(q(..))

end do

This allocation/deallocation pattern will result in deallocations from the
bottom of the asymmetric portion of the heap and allocations at the top of the
asymmetric portion of the heap. Consequentially, this will cut into the space
available for allocatable allocations.

This does the following to help address the issue:

- Adding uhcaf_print_alloc_slots routine for debugging

- Incorporate a linked list of empty slots into our managed RMA segment
  structure. Instead of first trying to split from the common slot, we from
  the first empty split opposite the common slot and search for an empty slot
  of sufficient size, working towards the common slot. The results in a more
  effective utilization of the space, but also might add some additional
  memory management overhead.

- Extended UHCAF_OUT_OF_SEGMENT_RMA so that if it is set to 2, then no
  asymmetric allocations from the RMA segment will be done. Not recommended to
  use this setting, however.

git-svn-id: http://svn.open64.net/svnroot/open64@4129 ead328b1-e903-0410-a19d-eacee4b031f6